Adjust Api models
