+++ src/Vendr.PaymentProviders.Dibs/Easy/Api/Models/DibsPaymentBase.cs
@@ -1,0 +1,16 @@
+﻿using Newtonsoft.Json;
+
+namespace Vendr.PaymentProviders.Dibs.Easy.Api.Models
+{
+    public class DibsPaymentBase
+    {
+        [JsonProperty("checkout")]
+        public DibsPaymentCheckout Checkout { get; set; }
+    }
+
+    public class DibsPaymentCheckout
+    {
+        [JsonProperty("url")]
+        public string Url { get; set; }
+    }
+}
