+++ src/Vendr.PaymentProviders.Dibs/Easy/Api/Models/DibsRefund.cs
@@ -1,0 +1,10 @@
+﻿using Newtonsoft.Json;
+
+namespace Vendr.PaymentProviders.Dibs.Easy.Api.Models
+{
+    public class DibsRefund
+    {
+        [JsonProperty("refundId")]
+        public string RefundId { get; set; }
+    }
+}
