+++ src/Vendr.PaymentProviders.Dibs/Vendr.PaymentProviders.Dibs.csproj
@@ -92,16 +92,20 @@
     <Compile Include="DibsSettingsBase.cs" />
     <Compile Include="DibsSettingsEasyBase.cs" />
     <Compile Include="Easy\Api\DibsEasyClient.cs" />
+    <Compile Include="Easy\Api\Models\DibsCancel.cs" />
     <Compile Include="Easy\Api\Models\DibsCharge.cs" />
     <Compile Include="Easy\Api\Models\DibsEasyClientConfig.cs" />
     <Compile Include="Easy\Api\Models\DibsOrder.cs" />
+    <Compile Include="Easy\Api\Models\DibsOrderItem.cs" />
     <Compile Include="Easy\Api\Models\DibsPayment.cs" />
+    <Compile Include="Easy\Api\Models\DibsPaymentBase.cs" />
     <Compile Include="Easy\Api\Models\DibsPaymentMethod.cs" />
     <Compile Include="Easy\Api\Models\DibsPaymentRequest.cs" />
     <Compile Include="Easy\Api\Models\DibsPaymentResult.cs" />
     <Compile Include="DibsLang.cs" />
     <Compile Include="DibsPaymentProvider.cs" />
     <Compile Include="DibsSettings.cs" />
+    <Compile Include="Easy\Api\Models\DibsRefund.cs" />
     <Compile Include="Easy\Api\Models\DibsWebhookEvent.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Properties\VersionInfo.cs" />
